requirements.txt
wheel
requests
click
numpy
tifffile
pillow
python-dateutil
utm
shapely
boto3
botocore
pyproj
oauthlib
requests_oauthlib
pyjwt
aenum>=2.1.4
